Clean up in lu. Added some algebra in qfunc. Small fix in convolution to support BigInt

+++ CBlock/cb.jl
@@ -335,8 +335,8 @@ function v_convCoeffs(sigma::Real,tup::(Int64,Int64)) #gives a dictionnary with 
                                                      # derivatives of u^sigma G
         m=tup[1] #a derivatives
         n=tup[2] #b derivatives
-		m=convert(Int128,m)
-		n=convert(Int128,n)
+		m=convert(BigInt,m)
+		n=convert(BigInt,n)
         d=sigma
 
         #comment: extra powers of 2 relative to my Mathematica code is due to the way the derivatives are defined by the Pythonists
@@ -351,8 +351,8 @@ function u_convCoeffs(sigma::Real,tup::(Int64,Int64)) #gives a dictionnary with 
                                                      # derivative (m,n) (in the tuple) of u^sigma G
         m=tup[1] #a derivatives
         n=tup[2] #b derivatives
-		m=convert(Int128,m)
-		n=convert(Int128,n)
+		m=convert(BigInt,m)
+		n=convert(BigInt,n)
         d=sigma
 
         #return [ (k,l)=> 1/(2^(k+2*l))*(-1)^(m-k)*binomial(n,n-l)*binomial(m,m-k)*(4)^(-d)*pochhammer(-2d+2*(n-l),m-k)*pochhammer(-d, n-l) for k=0:m, l=0:n]
